Add fallback mechanism for determining activity duration
Some games like L.A. Noire seem to switch application IDs after starting
the activity, which also makes the Discord API return a null start time
for the activity when it ends. To counter this, we'll store the latest
activity for the user in memory and use that if the normal start time is
not available once the activity ends.